% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footyr.R
\docType{package}
\name{footyr}
\alias{footyr}
\alias{footyr-package}
\title{footyr Package}
\description{
Allows users to easily access data from the ISport API through various functions.
Access data on over 1500 leagues/cups worldwide.
Get info on team standings, player statistics, league tables and more!
}
\section{footyr functions}{

footyr has a total of 5 functions as of version 0.1.0. These are
`ID`, `myleague`, `myplayer`, `myteam` and `mytopstats`. These
help aid in querying and parsing data from the ISport API and
presenting data in an easily readable format for the user.
You may obtain team, squad data, top statistics in league, and
other various football-related data using these functions
}

\author{
Timothy Young \email{timothyyoung@live.com.sg}
}
