% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{ID}
\alias{ID}
\title{Get ID numbers for football leagues and teams.}
\usage{
ID(league_name, team_name = NA)
}
\arguments{
\item{league_name}{(string) The name of the league/cup/competition that you
are looking to retrieve the ID number for. Required. No default.}

\item{team_name}{(string) The name of the team in the league that you
want to pull the team ID number for. Optional. Defaults to NA.}
}
\value{
A dataframe with league and team (optional) ID numbers.
}
\description{
This is a basic function that allows you to retrieve the ID numbers of
leagues/cups and teams.
}
\details{
This function serves as the basis for the rest of the functions in the
package. Upon input of strings for league and team names, the function outputs
the ID numbers of the league and teams, that you will need to use in other
functions in this package, as the API uses these ID numbers for queries.
}
\examples{
ID("English Premier League", "Arsenal")

ID("Spanish La Liga", "FC Barcelona")
}
\keyword{ID,}
\keyword{football,}
\keyword{league,}
\keyword{team}
