% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myleague.R
\name{myleague}
\alias{myleague}
\title{Get football league data and standings tables}
\usage{
myleague(leagueID, table = "no")
}
\arguments{
\item{leagueID}{The ID number associated with the league/cup as
defined by the API. This number can be obtained using the ID()
function also in this package. Required, no default value.}

\item{table}{(string) A binary parameter taking values "yes" or "no".
Will display the current league standings table if "yes" is chosen.
Optional. Defaults to "no".}
}
\value{
A dataframe with league/cup information and league table (optional).
}
\description{
This function that allows you to retrieve key information on
football leagues/cups, including the league standing tables.
}
\details{
This function pulls key statistics about a league/cup, including
current round, total rounds, current season. There is an optional argument
that will allow you to see the current league table that displays
the standings of all the current football teams in the league. The
table contains stats like number of wins, losses, draws, etc., ranking
the teams in terms of number of points gained in the league.
}
\examples{
myleague(1639, "yes")

myleague(188, "no")
}
\keyword{league,}
\keyword{table,}
\keyword{team}
