% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplayer.R
\name{myplayer}
\alias{myplayer}
\title{Get squad information of your football team}
\usage{
myplayer(teamID, player_name = NA)
}
\arguments{
\item{teamID}{The ID number associated with the team as
defined by the API. This number can be obtained using the ID()
function also in this package. Will output squad info
if this is only argument used. Required, no default value.}

\item{player_name}{(string) A parameter where you input the name
of the player whose data you want to view. This will output
a dataframe containing data about the chosen player only,
instead of displaying the entire squad info. Optional argument,
defaults to NA.}
}
\value{
A dataframe with team and player information.
}
\description{
This function that allows you to retrieve squad statistics of
a football team of your choice, including individual player and
coaching staff statistics.
}
\details{
This function pulls key data about a team's squad, including
full team list, squad numbers, players' birthdays and more. There
is one additional, optional argument that will allow you to
conduct a focussed search for an individual player within
that squad, which is useful if you know which player in particular
that you are searching for.
}
\examples{
myplayer(19, "Mesut Ozil")

myplayer(25)
}
\keyword{player}
\keyword{squad,}
\keyword{team,}
