% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myteam.R
\name{myteam}
\alias{myteam}
\title{Get detailed information on football team of your choice}
\usage{
myteam(leagueID, team_name, past_results = "no", future = NA)
}
\arguments{
\item{leagueID}{The ID number associated with the league/cup as
defined by the API. This number can be obtained using the ID()
function also in this package. Required, no default value.}

\item{team_name}{(string) A parameter where you input the name
of the team that you want to view. This will output a dataframe
containing data about their chosen football club. Required, no
default value for this argument.}

\item{past_results}{(string) A binary parameter taking values "yes" or "no".
Will display the past 6 match results that your team had played
if "yes" is chosen. Results are displayed in the last six columns
of the team data frame. Optional. Defaults to "no".}

\item{future}{An argument taking double-precision values. This argument
specifies the number of future fixtures that you would like to view.
Eg. 5 will show you the 5 upcoming matches for your team, for both home
and away games. Optional argument, defaults to NA.}
}
\value{
A dataframe with team information and past and future match results/fixtures.
}
\description{
This function that allows you to retrieve key statistics on
a football team of your choice, including past results and
future fixtures.
}
\details{
This function pulls key statistics about a team, including
founding date, manager name, stadium name and capacity. There
are two additional, optional arguments that will allow you to
toggle if you would like to view past and/or future fixtures.
You would be able to view the past 6 results from matches
your team played, and you can also specify the number of
future fixtures that you would like to view. In particular, the
team_name argument in this function will match the name
with the team ID that will be then used to query the API, thus
saving the user a step from looking up the ID for their desired
team.
}
\examples{
myteam(1639, "Arsenal", "yes", 5)

myteam(1134, "Real Madrid", "no", 2)
}
\keyword{fixtures,}
\keyword{league,}
\keyword{results,}
\keyword{team}
