% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytopstats.R
\name{mytopstats}
\alias{mytopstats}
\title{Get data and statistics of leagues' best players}
\usage{
mytopstats(leagueID, stat, more_details = "no")
}
\arguments{
\item{leagueID}{The ID number associated with the league/cup as
defined by the API. This number can be obtained using the ID()
function also in this package. Argument required, no default value.}

\item{stat}{(string) A parameter where you input the stat
that you would like to predominantly focus on. The chosen stat
will result in the dataframe output to be ordered based
on that stat. Available stats would include: "Goals",
"Shots", "Shots_on_Target", "Assists", "Key_Passes". This
argument is required, with no default.}

\item{more_details}{(string) A binary parameter taking
values "yes" or "no". If argument input is "yes",
additional stats would be included in the dataframe output,
including shot accuracy, mins per goal, etc. Argument is
optional, with default set to "no".}
}
\value{
A dataframe with players with the top statistics in the league.
}
\description{
This function that allows you to retrieve complete statistics of
the top players within a league, including additional stats not
natively from the ISports API.
}
\details{
This function pulls key statistics about the top players within
the league. This could include top goalscorers, most assists, most
shots, etc. The user simply specifies the league using the leagueID
as the argument, followed by the statistic/key variable that
they would like to focus on, and input this as the second argument
in the function. The user could also choose to view additional
statistics if they so desire by using the third binary argument.
}
\examples{
mytopstats(1639, "Goals", "yes")

mytopstats(188, "Assists")
}
\keyword{assists}
\keyword{goals,}
\keyword{league,}
\keyword{stats,}
\keyword{top,}
